backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.30
aiosqlite==0.20.0
httpx==0.27.0
python-dotenv==1.0.1
scikit-learn
nltk
pandas
numpy
joblib
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
